requirements.txt
amqp==5.0.2
anyio==3.4.0
appnope==0.1.2
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-lru==1.0.2
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bidict==0.21.4
billiard==3.6.3.0
bleach==4.1.0
celery==5.0.5
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
configparser==5.2.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2
django-environ==0.8.1
django-extensions==3.1.0
django-import-export==2.5.0
django-jsonfield==1.4.1
django-redis==5.2.0
django-simple-history==2.12.0
entrypoints==0.3
et-xmlfile==1.0.1
executing==0.8.2
fxcmpy==1.3.2
icecream==2.1.1
idna==3.3
importlib-metadata==3.3.0
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jdcal==1.4.1
jedi==0.18.1
Jinja2==3.0.3
json5==0.9.6
jsonschema==4.3.2
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-server==1.13.1
jupyterlab==3.2.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.1
jupyterlab-widgets==1.0.2
kombu==5.0.2
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
numpy==1.21.5
odfpy==1.4.1
openpyxl==3.0.5
packaging==21.3
pandas==1.3.5
pandas-ta==0.3.14b0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psycopg2==2.8.6
ptyprocess==0.7.0
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
Pygments==2.10.0
pyparsing==3.0.6
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-engineio==4.3.1
python-socketio==5.0.4
pytz==2021.3
PyYAML==5.3.1
pyzmq==22.3.0
redis==3.5.3
requests==2.26.0
rsa==4.7
schedule==1.1.0
Send2Trash==1.8.0
sentry-sdk==0.19.5
six==1.16.0
sniffio==1.2.0
socketIO-client==0.7.2
sqlparse==0.4.1
ta==0.8.0
tablib==3.0.0
Telethon==1.19.0
terminado==0.12.1
testpath==0.5.0
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
typing-extensions==3.7.4.3
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
widgetsnbextension==3.5.2
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
